Use suitable api version for OSC.
In manilaclient i.e. manila commands, if user supplied version is above
the max version supported by server, manilaclient adjust to max version.
This was missing in openstack commands which results in failure that
version is not supported. Fix by discovering most suitable version and
use it for client creation.

Closes-bug: #1960490
Change-Id: I734f6953e8b0266b38f0cb6e1581c4201f2bd676